Deleted diff model, going to do the same for GANs
The issue is that these both are surface level abstractions around the
proccess and the logic of these two should really be encapsulated by
their runner.

The one nice thing that we would lose is the call to `inference` being
stored in the nn.Module when loaded from disk.

Maybe we can add a hook to module before saving to get around this.